+++ src/constants/index.ts
@@ -138,5 +138,7 @@ export const COUNT_PER_PAGE = 5;
 export const COUNT_PER_SECTION = 5;
 export const RENT_DETAIL_COUNT_PER_PAGE = 3;
 export const RENT_DETAIL_COUNT_PER_SECTION = 5;
+export const ADMIN_RENT_LIST_COUNT_PER_PAGE = 10;
+export const ADMIN_RENT_LIST_COUNT_PER_SECTION = 5;
