+++ src/views/service/MyPage/MyPageRentList/UserRentList/style.css
@@ -1,0 +1,108 @@
+#user-rent-list-wrapper{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    
+    height: 100%;
+
+    padding: 40px 100px;
+    gap: 20px;
+
+    background-color: rgba(241,241,241,1);
+}
+
+.user-rent-list-table{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    background-color: rgba(255,255,255,1);
+    padding: 20px;
+    border-radius: 20px;
+
+    width: 400px;
+    height: 150px;
+}
+
+.user-rent-list-content{
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    width: 100%;
+    gap: 10px;
+}
+
+.user-rent-list-name{
+    font-family: "Noto Sans KR";
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+
+    color: rgba(0,0,0,1);
+}
+
+.user-rent-list-date{
+    display: flex;
+    flex-direction: row;
+    gap: 16px;
+
+    color: #666;
+    font-family: "Noto Sans KR";
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+}
+
+.user-rent-list-total-price, .user-rent-list-detail{
+    font-family: "Noto Sans KR";
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+
+    color: rgba(0,0,0,1);
+}
+
+.user-rent-list-page-button{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 3px;
+
+    cursor: pointer;
+}
+
+.user-rent-list-next-button{
+    width: 4px;
+    height: 8px;
+
+    background-image: url('../../../../../assets/image/rent-detail-button-image.png');
+    background-size: 100% 100%;
+}
+
+.user-rent-list-table-bottom{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap : 290px;
+}
+
+.user-rent-list-cancel-button {
+    padding: 5px 10px;
+    border: 1px solid rgba(0,0,0,1);    
+
+    font-family: "Noto Sans KR";
+    font-size: 13px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 175%;
+    text-align: center;
+}
+
+.user-rent-list-cancel-button:hover{
+    cursor: pointer;
+    background-color: rgba(217,217,217,1);
+}
