+++ src/views/service/MyPage/MyPageRentList/style.css
@@ -1,129 +1,3 @@
 #mp-rent-list-wrapper{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    
     height: 100%;
-
-    padding: 40px 100px;
-    gap: 20px;
-
-    background-color: rgba(241,241,241,1);
-}
-
-.mp-rent-list-table{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-    background-color: rgba(255,255,255,1);
-    padding: 20px;
-    border-radius: 20px;
-
-    width: 400px;
-    height: 150px;
-}
-
-.mp-rent-list-content{
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    width: 100%;
-    gap: 10px;
-}
-
-.mp-rent-list-name{
-    font-family: "Noto Sans KR";
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-
-    color: rgba(0,0,0,1);
-}
-
-.mp-rent-list-date{
-    display: flex;
-    flex-direction: row;
-    gap: 16px;
-
-    color: #666;
-    font-family: "Noto Sans KR";
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-}
-
-.mp-rent-list-total-price, .mp-rent-list-detail{
-    font-family: "Noto Sans KR";
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-
-    color: rgba(0,0,0,1);
-}
-
-.mp-rent-list-page-button{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 3px;
-
-    cursor: pointer;
-}
-
-.mp-rent-list-next-button{
-    width: 4px;
-    height: 8px;
-
-    background-image: url('../../../../assets/image/rent-detail-button-image.png');
-    background-size: 100% 100%;
-}
-
-.mp-rent-list-table-bottom{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap : 290px;
-}
-
-.mp-rent-list-cancel-button {
-    margin-top: 10px;
-    padding: 5px 10px;
-    background-color: red;
-    color: white;
-    cursor: pointer;
-
-    font-family: "Noto Sans KR";
-    font-size: 13px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 175%;
-    text-align: center;
-}
-
-.mp-rent-list-status-change {
-    display: flex;
-    margin-top: 10px;
-}
-
-.mp-rent-list-status-change select {
-    width: 80px;
-    height: 20px; 
-    margin-bottom: 5px;
-    padding: 0;
-}
-
-.mp-rent-list-status-change button {
-    width: 80px;
-    height: 20px; 
-    padding: 0;
-    background-color: blue;
-    color: white;
-    cursor: pointer;
-    border: none;
-    border-radius: 4px;
 }
